fix: prevent subtraction overflow in attestation size_reduction calc
When fused output is larger than input (e.g., single-component fusion),
`input_size - output_size` overflows on unsigned integers. Convert to
f64 before subtracting so the result can be negative.

+++ meld-core/src/lib.rs
@@ -772,7 +772,7 @@ impl Fuser {
             serde_json::json!(stats.imports_resolved),
         );
         let size_reduction = if stats.input_size > 0 {
-            ((stats.input_size - stats.output_size) as f64 / stats.input_size as f64) * 100.0
+            ((stats.input_size as f64 - stats.output_size as f64) / stats.input_size as f64) * 100.0
         } else {
             0.0
         };
